Added my Part of ChallengeViewController
1. added methods: answerCorrect, answerIncorrect, and
(IBAction)nextProblem:(id)sender
++These methods have been tested. I modified them so they were
(IBAction)method:(id)sender and attached them to the Next and Submit
buttons. Score, Alert pop-up, and nextProblem up to 10 have all been
tested.

2. added to ModeViewController...
I added the UIAlertView to the MVC so, that when 10 problems is
reached, an alert pops-up, and tells the user their score. The working
title is currently "Alert!" which I HATE. Someone should be given the
job of making the continue button go to the home screen on the 10
problems alert. It'd also probably be best to make the other AlertViews
hook up to this one.

3. Did not do commenting yet/final analysis of names.